Revert "fix: course fetching, filtering, sorting, and pagination in MyCourses…"
This reverts commit b5cee29d27e886a2f133a9e55d4d19a49dea7e81.
